refactor(ledger): remove unreachable adoption path
