+++ internal/ledger/service/fix_mismatch_test.go
@@ -1,7 +1,7 @@
 package service
 
 import (
-	"context"
+	"strconv"
 	"testing"
 
 	"github.com/LeJamon/go-xrpl/drops"
@@ -47,7 +47,7 @@ func makeStubStateMap(t *testing.T, seq uint32) (*shamap.SHAMap, [32]byte) {
 	return stateMap, root
 }
 
-// TestAdoptLedgerWithState_FixMismatchInvalidatesDivergedTail pins F5:
+// TestSwitchToPreferredLedger_FixMismatchInvalidatesDivergedTail pins F5:
 // when the adopted ledger's parent hash does NOT match the hash of the
 // prev-seq entry in ledgerHistory, the diverged slot (and every
 // orphaned forward entry) must be purged before the adopt installs
@@ -64,7 +64,7 @@ func makeStubStateMap(t *testing.T, seq uint32) (*shamap.SHAMap, [32]byte) {
 //
 // Regression guard: before F5, the old B3 would be silently overwritten
 // and C3 would linger as an orphan on the wrong fork.
-func TestAdoptLedgerWithState_FixMismatchInvalidatesDivergedTail(t *testing.T) {
+func TestSwitchToPreferredLedger_FixMismatchInvalidatesDivergedTail(t *testing.T) {
 	cfg := DefaultConfig()
 	svc, err := New(cfg)
 	require.NoError(t, err)
@@ -85,7 +85,7 @@ func TestAdoptLedgerWithState_FixMismatchInvalidatesDivergedTail(t *testing.T) {
 	ledB := makeStubLedger(t, baseSeq+1, hashB, hashA) // B chains to A
 	ledC := makeStubLedger(t, baseSeq+2, hashC, hashB) // C chains to B
 
-	// Seed the in-memory history directly. Bypass AdoptLedgerWithState
+	// Seed the in-memory history directly. Bypass SwitchToPreferredLedger
 	// to isolate the F5 path under test.
 	svc.mu.Lock()
 	svc.ledgerHistory[ledA.Sequence()] = ledA
@@ -105,25 +105,8 @@ func TestAdoptLedgerWithState_FixMismatchInvalidatesDivergedTail(t *testing.T) {
 	hashD[0] = 0xDD
 	var divergentParent [32]byte
 	divergentParent[0] = 0xFF // deliberately != hashA
-
-	stateMap := shamap.New(shamap.TypeState)
-	stateRoot, err := stateMap.Hash()
-	require.NoError(t, err)
-
-	txMap := shamap.New(shamap.TypeTransaction)
-	txRoot, err := txMap.Hash()
-	require.NoError(t, err)
-
-	hdrD := &header.LedgerHeader{
-		LedgerIndex: baseSeq + 1,
-		ParentHash:  divergentParent,
-		TxHash:      txRoot,
-		AccountHash: stateRoot,
-	}
-	hdrD.Hash = header.CalculateHash(*hdrD)
-	hashD = hdrD.Hash
-
-	require.NoError(t, svc.AdoptLedgerWithState(context.TODO(), hdrD, stateMap, txMap))
+	preferred := makeStubLedger(t, baseSeq+1, hashD, divergentParent)
+	require.NoError(t, svc.SwitchToPreferredLedger(preferred))
 
 	svc.mu.RLock()
 	defer svc.mu.RUnlock()
@@ -145,17 +128,17 @@ func TestAdoptLedgerWithState_FixMismatchInvalidatesDivergedTail(t *testing.T) {
 	require.NotNil(t, svc.closedLedger)
 	assert.Equal(t, hashD, svc.closedLedger.Hash(),
 		"closedLedger must track the adopted ledger after a fork-switch adopt")
-	assert.Equal(t, "empty", svc.completeLedgersString(),
-		"fork invalidation must remove every purged ledger from complete_ledgers")
+	assert.Equal(t, strconv.FormatUint(uint64(baseSeq+1), 10), svc.completeLedgersString(),
+		"fork invalidation must retain only the preferred ledger in complete_ledgers")
 }
 
-// TestAdoptLedgerWithState_NoMismatchNoOp pins the happy path:
+// TestSwitchToPreferredLedger_NoMismatchNoOp pins the happy path:
 // if the adopted ledger's parentHash matches the hash of the prev-seq
 // entry in ledgerHistory, fixMismatch must leave history alone.
 //
 // Seed: A@S, B@S+1 (child of A). Adopt D@S+2 whose parentHash == B.Hash().
 // After adopt: A, B, D must all remain in history, D is new.
-func TestAdoptLedgerWithState_NoMismatchNoOp(t *testing.T) {
+func TestSwitchToPreferredLedger_NoMismatchNoOp(t *testing.T) {
 	cfg := DefaultConfig()
 	svc, err := New(cfg)
 	require.NoError(t, err)
@@ -180,23 +163,8 @@ func TestAdoptLedgerWithState_NoMismatchNoOp(t *testing.T) {
 	var hashD [32]byte
 	hashD[0] = 0xD1
 
-	stateMap := shamap.New(shamap.TypeState)
-	stateRoot, err := stateMap.Hash()
-	require.NoError(t, err)
-	txMap := shamap.New(shamap.TypeTransaction)
-	txRoot, err := txMap.Hash()
-	require.NoError(t, err)
-
-	hdrD := &header.LedgerHeader{
-		LedgerIndex: baseSeq + 2,
-		ParentHash:  hashB, // chains correctly to B
-		TxHash:      txRoot,
-		AccountHash: stateRoot,
-	}
-	hdrD.Hash = header.CalculateHash(*hdrD)
-	hashD = hdrD.Hash
-
-	require.NoError(t, svc.AdoptLedgerWithState(context.TODO(), hdrD, stateMap, txMap))
+	preferred := makeStubLedger(t, baseSeq+2, hashD, hashB)
+	require.NoError(t, svc.SwitchToPreferredLedger(preferred))
 
 	svc.mu.RLock()
 	defer svc.mu.RUnlock()
@@ -215,82 +183,7 @@ func TestAdoptLedgerWithState_NoMismatchNoOp(t *testing.T) {
 	assert.Equal(t, hashD, gotD.Hash())
 }
 
-// TestAdoptLedgerWithState_BackfillAtForkBoundaryKeepsCanonicalChain pins the
-// below-tip guard in fixMismatch: a backfilled ledger whose parent mismatches
-// the stale fork entry below it but whose canonical entry above chains to it
-// must purge only the fork ledger, never the canonical chain above (which the
-// general mismatch path would sweep as "orphans", nuking the adopted tip).
-func TestAdoptLedgerWithState_BackfillAtForkBoundaryKeepsCanonicalChain(t *testing.T) {
-	cfg := DefaultConfig()
-	svc, err := New(cfg)
-	require.NoError(t, err)
-	require.NoError(t, svc.Start())
-
-	baseSeq := svc.GetClosedLedgerIndex() + 1
-
-	var hashFork, hashC3, hashC4, hashC5, hashC2 [32]byte
-	hashFork[0] = 0xF0 // stale fork ledger at baseSeq
-	hashC2[0] = 0xC2   // canonical parent of C3 (not in history)
-	hashC3[0] = 0xC3   // the backfilled ledger at baseSeq+1
-	hashC4[0] = 0xC4   // canonical at baseSeq+2, chains to C3
-	hashC5[0] = 0xC5   // canonical closed tip at baseSeq+3
-
-	var zero [32]byte
-	fork := makeStubLedger(t, baseSeq, hashFork, zero)
-
-	stateMap := shamap.New(shamap.TypeState)
-	stateRoot, err := stateMap.Hash()
-	require.NoError(t, err)
-	txMap := shamap.New(shamap.TypeTransaction)
-	txRoot, err := txMap.Hash()
-	require.NoError(t, err)
-
-	// The backfilled C3: parent C2 != fork.Hash (fork boundary below), but
-	// C4.ParentHash == C3.Hash (canonical above).
-	hdrC3 := &header.LedgerHeader{
-		LedgerIndex: baseSeq + 1,
-		ParentHash:  hashC2,
-		TxHash:      txRoot,
-		AccountHash: stateRoot,
-	}
-	hdrC3.Hash = header.CalculateHash(*hdrC3)
-	hashC3 = hdrC3.Hash
-	c4 := makeStubLedger(t, baseSeq+2, hashC4, hashC3)
-	c5 := makeStubLedger(t, baseSeq+3, hashC5, hashC4)
-
-	svc.mu.Lock()
-	svc.ledgerHistory[fork.Sequence()] = fork
-	svc.ledgerHistory[c4.Sequence()] = c4
-	svc.ledgerHistory[c5.Sequence()] = c5
-	svc.closedLedger = c5
-	svc.mu.Unlock()
-
-	require.NoError(t, svc.AdoptLedgerWithState(context.TODO(), hdrC3, stateMap, txMap))
-
-	svc.mu.RLock()
-	defer svc.mu.RUnlock()
-
-	gotC3, okC3 := svc.ledgerHistory[baseSeq+1]
-	require.True(t, okC3, "backfilled C3 must be installed")
-	assert.Equal(t, hashC3, gotC3.Hash())
-
-	_, okFork := svc.ledgerHistory[baseSeq]
-	assert.False(t, okFork, "the stale fork ledger below the backfill must be purged")
-
-	gotC4, okC4 := svc.ledgerHistory[baseSeq+2]
-	require.True(t, okC4, "canonical C4 above the backfill must survive")
-	assert.Equal(t, hashC4, gotC4.Hash())
-
-	gotC5, okC5 := svc.ledgerHistory[baseSeq+3]
-	require.True(t, okC5, "canonical closed tip C5 must survive")
-	assert.Equal(t, hashC5, gotC5.Hash())
-
-	require.NotNil(t, svc.closedLedger)
-	assert.Equal(t, hashC5, svc.closedLedger.Hash(),
-		"a below-tip backfill must not move or clear the closed pointer")
-}
-
-// TestAdoptLedgerWithState_FixMismatchValidatedLedgerInvalidationLogsError
+// TestSwitchToPreferredLedger_FixMismatchValidatedLedgerInvalidationLogsError
 // pins the escalation behavior: if fixMismatch invalidates a ledger that
 // was already quorum-validated, it MUST NOT silently reset
 // s.validatedLedger. A validated-ledger invalidation indicates a genuine
@@ -301,7 +194,7 @@ func TestAdoptLedgerWithState_BackfillAtForkBoundaryKeepsCanonicalChain(t *testi
 // prior pointer after fixMismatch runs on a validated-prev case.
 // The ERROR log itself is observed through the exercised code path;
 // verifying the log message text is brittle and is not checked here.
-func TestAdoptLedgerWithState_FixMismatchValidatedLedgerInvalidationLogsError(t *testing.T) {
+func TestSwitchToPreferredLedger_FixMismatchValidatedLedgerInvalidationLogsError(t *testing.T) {
 	cfg := DefaultConfig()
 	svc, err := New(cfg)
 	require.NoError(t, err)
@@ -310,7 +203,7 @@ func TestAdoptLedgerWithState_FixMismatchValidatedLedgerInvalidationLogsError(t 
 	baseSeq := svc.GetClosedLedgerIndex() + 1
 
 	// Seed B at seq S+0; makeStubLedger explicitly marks the header
-	// validated. Adopt D at seq
+	// validated. Switch to D at seq
 	// S+1 whose parentHash does not equal B.Hash() — fixMismatch must
 	// purge B and log ERROR. The validatedLedger pointer must not
 	// silently flip.
@@ -329,23 +222,10 @@ func TestAdoptLedgerWithState_FixMismatchValidatedLedgerInvalidationLogsError(t 
 
 	var divergentParent [32]byte
 	divergentParent[0] = 0xAB
-
-	stateMap := shamap.New(shamap.TypeState)
-	stateRoot, err := stateMap.Hash()
-	require.NoError(t, err)
-	txMap := shamap.New(shamap.TypeTransaction)
-	txRoot, err := txMap.Hash()
-	require.NoError(t, err)
-
-	hdrD := &header.LedgerHeader{
-		LedgerIndex: baseSeq + 1,
-		ParentHash:  divergentParent,
-		TxHash:      txRoot,
-		AccountHash: stateRoot,
-	}
-	hdrD.Hash = header.CalculateHash(*hdrD)
-
-	require.NoError(t, svc.AdoptLedgerWithState(context.TODO(), hdrD, stateMap, txMap))
+	var hashD [32]byte
+	hashD[0] = 0xDD
+	preferred := makeStubLedger(t, baseSeq+1, hashD, divergentParent)
+	require.NoError(t, svc.SwitchToPreferredLedger(preferred))
 
 	svc.mu.RLock()
 	defer svc.mu.RUnlock()
@@ -355,11 +235,11 @@ func TestAdoptLedgerWithState_FixMismatchValidatedLedgerInvalidationLogsError(t 
 			"requires operator action, not silent rewrite")
 }
 
-// TestAdoptLedgerWithState_FixMismatchPurgesTxIndex pins that when a
+// TestSwitchToPreferredLedger_FixMismatchPurgesTxIndex pins that when a
 // ledger is invalidated by fixMismatch, any tx-index entries pointing
 // at it are removed too. Otherwise `tx` RPCs would keep resolving tx
 // hashes to a ledger slot whose contents were just discarded.
-func TestAdoptLedgerWithState_FixMismatchPurgesTxIndex(t *testing.T) {
+func TestSwitchToPreferredLedger_FixMismatchPurgesTxIndex(t *testing.T) {
 	cfg := DefaultConfig()
 	svc, err := New(cfg)
 	require.NoError(t, err)
@@ -395,23 +275,10 @@ func TestAdoptLedgerWithState_FixMismatchPurgesTxIndex(t *testing.T) {
 	// Divergent D at seq S+1.
 	var divergentParent [32]byte
 	divergentParent[0] = 0xEE
-
-	stateMap := shamap.New(shamap.TypeState)
-	stateRoot, err := stateMap.Hash()
-	require.NoError(t, err)
-	txMap := shamap.New(shamap.TypeTransaction)
-	txRoot, err := txMap.Hash()
-	require.NoError(t, err)
-
-	hdrD := &header.LedgerHeader{
-		LedgerIndex: baseSeq + 1,
-		ParentHash:  divergentParent,
-		TxHash:      txRoot,
-		AccountHash: stateRoot,
-	}
-	hdrD.Hash = header.CalculateHash(*hdrD)
-
-	require.NoError(t, svc.AdoptLedgerWithState(context.TODO(), hdrD, stateMap, txMap))
+	var hashD [32]byte
+	hashD[0] = 0xDD
+	preferred := makeStubLedger(t, baseSeq+1, hashD, divergentParent)
+	require.NoError(t, svc.SwitchToPreferredLedger(preferred))
 
 	svc.mu.RLock()
 	defer svc.mu.RUnlock()
