+++ internal/ledger/service/lifecycle.go
@@ -220,27 +220,6 @@ func ledgerInfo(closed *ledger.Ledger) *LedgerInfo {
 	}
 }
 
-// installAdoptedLedgerLocked writes adopted into ledgerHistory[seq] under the
-// validated-precedence rule and returns the canonical entry; callers must use the
-// return as s.closedLedger to keep history and closed-reference consistent.
-// Caller must hold s.mu (write).
-func (s *Service) installAdoptedLedgerLocked(seq uint32, adopted *ledger.Ledger) *ledger.Ledger {
-	if existing, ok := s.ledgerHistory[seq]; ok {
-		existingHash := existing.Hash()
-		newHash := adopted.Hash()
-		if existingHash != newHash && existing.IsValidated() && !adopted.IsValidated() {
-			s.logger.Warn("adopt skip: validated entry already present",
-				"seq", seq,
-				"existing_hash", fmt.Sprintf("%x", existingHash[:8]),
-				"adopt_hash", fmt.Sprintf("%x", newHash[:8]),
-			)
-			return existing
-		}
-	}
-	s.putHistoryLocked(adopted)
-	return adopted
-}
-
 // fixMismatchLocked invalidates the tail of ledgerHistory when adopted does not
 // chain to the entry at adopted.Sequence()-1. On mismatch it purges the prev-seq
 // slot and every seq > adoptedSeq (orphaned forward entries), drops their
@@ -958,139 +937,6 @@ func (s *Service) setValidatedLedgerAt(seq uint32, expectedHash [32]byte, signTi
 	notification.notify()
 }
 
-// SubmitHeldAdoptionResult describes the disposition of a candidate ledger. When
-// Stashed, the caller must arm a backward acquisition for (ParentSeq, ParentHash)
-// or the entry ages out at heldAdoptionTTL (issue #397).
-type SubmitHeldAdoptionResult struct {
-	// Adopted: the awaited parent was already in history at the expected hash.
-	Adopted bool
-
-	// Stashed: parked in the held-adoption stash pending cascade at the parent seq.
-	Stashed bool
-
-	// ParentSeq, ParentHash describe the awaited parent. Set whenever
-	// h.LedgerIndex > 1, regardless of outcome.
-	ParentSeq  uint32
-	ParentHash [32]byte
-}
-
-// SubmitHeldAdoption routes a fetched replay-delta either to immediate adoption
-// (awaited parent already in history at the matching hash) or to the held-orphan
-// stash keyed by the awaited parent seq, cascade-adopted later from
-// AdoptLedgerWithState. Safe to call concurrently. Nil header/stateMap rejected;
-// nil txMap is allowed (legacy catchup → empty genesis-shaped tx map).
-func (s *Service) SubmitHeldAdoption(ctx context.Context, h *header.LedgerHeader, stateMap *shamap.SHAMap, txMap *shamap.SHAMap) (SubmitHeldAdoptionResult, error) {
-	if h == nil {
-		return SubmitHeldAdoptionResult{}, errors.New("SubmitHeldAdoption: nil header")
-	}
-	if stateMap == nil {
-		return SubmitHeldAdoptionResult{}, errors.New("SubmitHeldAdoption: nil state map")
-	}
-
-	res := SubmitHeldAdoptionResult{}
-	if h.LedgerIndex > 1 {
-		res.ParentSeq = h.LedgerIndex - 1
-		res.ParentHash = h.ParentHash
-	}
-
-	s.mu.Lock()
-	previousValidated := s.validatedLedger
-	defer s.unlockAndNotifyValidatedLedger(previousValidated)
-	s.historyComponent.mu.Lock()
-	defer s.historyComponent.mu.Unlock()
-
-	if h.LedgerIndex == 0 {
-		return res, errors.New("SubmitHeldAdoption: ledger sequence must be non-zero")
-	}
-	if _, err := s.ledgerWithStateLocked(h, stateMap, txMap); err != nil {
-		return res, fmt.Errorf("SubmitHeldAdoption: validate candidate: %w", err)
-	}
-
-	// Evict stale entries on every submission.
-	s.evictExpiredHeldAdoptionsLocked()
-
-	// Fast path: parent already in history at the expected hash → adopt now
-	// (skipped for seq <= 1, which has no parent).
-	if h.LedgerIndex > 1 {
-		parentSeq := h.LedgerIndex - 1
-		if parent, ok := s.ledgerHistory[parentSeq]; ok {
-			parentHash := parent.Hash()
-			if parentHash == h.ParentHash {
-				if err := s.adoptLedgerWithStateLocked(ctx, h, stateMap, txMap, 0); err != nil {
-					return res, err
-				}
-				res.Adopted = true
-				return res, nil
-			}
-			// Parent present on a different fork — stash; cascade adopts once the
-			// awaited parent arrives and fixMismatchLocked clears the tail.
-			s.logger.Info("SubmitHeldAdoption divergent-parent submission stashed",
-				"seq", h.LedgerIndex,
-				"parent_seq", parentSeq,
-				"parent_have", fmt.Sprintf("%x", parentHash[:8]),
-				"parent_want", fmt.Sprintf("%x", h.ParentHash[:8]),
-			)
-		}
-	}
-
-	// Parent not yet present — stash.
-	s.heldAdoptions[h.LedgerIndex-1] = &pendingAdopt{
-		header:   h,
-		stateMap: stateMap,
-		txMap:    txMap,
-		at:       time.Now(),
-	}
-	res.Stashed = true
-	return res, nil
-}
-
-// cascadeHeldAdoptionsLocked promotes the held child awaiting the just-adopted
-// seq when its ParentHash matches the adopted hash, recursing through any chain
-// of pre-stashed orphans (bounded by heldAdoptionCascadeMax). Evicts entries
-// older than heldAdoptionTTL on every call. Caller must hold s.mu (write).
-func (s *Service) cascadeHeldAdoptionsLocked(ctx context.Context, adopted *ledger.Ledger, depth int) {
-	s.evictExpiredHeldAdoptionsLocked()
-
-	if depth >= heldAdoptionCascadeMax {
-		s.logger.Warn("cascadeHeldAdoptions: hit recursion cap — refusing further promotion",
-			"cap", heldAdoptionCascadeMax,
-			"seq", adopted.Sequence(),
-		)
-		return
-	}
-
-	parentSeq := adopted.Sequence()
-	held, ok := s.heldAdoptions[parentSeq]
-	if !ok {
-		return
-	}
-	delete(s.heldAdoptions, parentSeq)
-
-	adoptedHash := adopted.Hash()
-	if held.header.ParentHash != adoptedHash {
-		// Held orphan expected a different parent hash — divergent fork, drop it.
-		s.logger.Warn("cascadeHeldAdoptions: dropping fork-mismatched held entry",
-			"seq", held.header.LedgerIndex,
-			"parent_have", fmt.Sprintf("%x", adoptedHash[:8]),
-			"parent_want", fmt.Sprintf("%x", held.header.ParentHash[:8]),
-		)
-		return
-	}
-
-	s.logger.Info("cascadeHeldAdoptions: promoting held orphan",
-		"seq", held.header.LedgerIndex,
-		"hash", fmt.Sprintf("%x", held.header.Hash[:8]),
-		"depth", depth+1,
-	)
-	if err := s.adoptLedgerWithStateLocked(ctx, held.header, held.stateMap, held.txMap, depth+1); err != nil {
-		// Cascade-hop adopt failed; log and stop — the outer adopt succeeded.
-		s.logger.Error("cascadeHeldAdoptions: held-entry adopt failed",
-			"seq", held.header.LedgerIndex,
-			"err", err,
-		)
-	}
-}
-
 func (s *Service) completeInitialSyncLocked() {
 	if s.networkLedgerState == networkLedgerNeeded {
 		s.networkLedgerState = networkLedgerReady
@@ -1218,18 +1064,6 @@ func (s *Service) IngestHistoricalLedgerWithState(ctx context.Context, h *header
 	return nil
 }
 
-// AdoptLedgerWithState adopts a ledger using a fully-fetched state map from a
-// peer. txMap is the verified tx SHAMap on the replay-delta path; pass nil for
-// state-only catchup of a ledger whose transaction root is empty.
-func (s *Service) AdoptLedgerWithState(ctx context.Context, h *header.LedgerHeader, stateMap *shamap.SHAMap, txMap *shamap.SHAMap) error {
-	s.mu.Lock()
-	previousValidated := s.validatedLedger
-	defer s.unlockAndNotifyValidatedLedger(previousValidated)
-	s.historyComponent.mu.Lock()
-	defer s.historyComponent.mu.Unlock()
-	return s.adoptLedgerWithStateLocked(ctx, h, stateMap, txMap, 0)
-}
-
 func (s *Service) ledgerWithStateLocked(h *header.LedgerHeader, stateMap *shamap.SHAMap, txMap *shamap.SHAMap) (*ledger.Ledger, error) {
 	if s.genesisLedger == nil {
 		return nil, errors.New("no genesis ledger available")
@@ -1292,124 +1126,5 @@ func (s *Service) storeLedgerWithStateLocked(ctx context.Context, h *header.Ledg
 		"seq", stored.Sequence(),
 		"hash", fmt.Sprintf("%x", storedHash[:8]),
 	)
-	return nil
-}
-
-// adoptLedgerWithStateLocked is the lock-held core of AdoptLedgerWithState.
-// Caller must hold s.mu (write). cascadeDepth is the held-orphan cascade depth;
-// public entrypoints pass 0, the cascade recurses to heldAdoptionCascadeMax.
-func (s *Service) adoptLedgerWithStateLocked(
-	ctx context.Context,
-	h *header.LedgerHeader,
-	stateMap *shamap.SHAMap,
-	txMap *shamap.SHAMap,
-	cascadeDepth int,
-) error {
-	adopted, err := s.ledgerWithStateLocked(h, stateMap, txMap)
-	if err != nil {
-		return err
-	}
-	stagedResults, err := stageTransactionResults(adopted, h.LedgerIndex, h.Hash)
-	if err != nil {
-		return fmt.Errorf("collect adopted transaction results: %w", err)
-	}
-
-	// Invalidate the history tail if adopted doesn't chain to our seq-1 entry
-	// (divergent fork) so RPCs don't resolve stale state. See fixMismatchLocked.
-	s.fixMismatchLocked(adopted)
-
-	// Install into history; advance closedLedger only on strict seq increase
-	// (backward cascade fills must not regress it).
-	canonical := s.installAdoptedLedgerLocked(h.LedgerIndex, adopted)
-	advanced := false
-	if s.closedLedger == nil || canonical.Sequence() > s.closedLedger.Sequence() {
-		s.closedLedger = canonical
-		advanced = true
-	} else if canonical.Sequence() == s.closedLedger.Sequence() && canonical.Hash() != s.closedLedger.Hash() {
-		// Sibling-fork (issue #470): a same-seq adopt with a different hash
-		// replaces our local alt; closedLedger must point at adopted or later
-		// builds keep snapshotting the alt's state map and diverge forever.
-		s.closedLedger = canonical
-		advanced = true
-	}
-	s.completeInitialSyncLocked()
-
-	// The canonical entry has already completed persistence, validation
-	// draining, and event collection.
-	if canonical != adopted {
-		openLedger, err := ledger.NewOpen(canonical, time.Now())
-		if err != nil {
-			return fmt.Errorf("failed to create open ledger after adopt-skip: %w", err)
-		}
-		s.openLedger = openLedger
-		if advanced {
-			if err := s.rebuildOpenLedgerViewLocked(); err != nil {
-				return err
-			}
-		}
-		canonicalHash := canonical.Hash()
-		s.logger.Info("Adopted ledger from peer (skip: validated entry kept)",
-			"seq", h.LedgerIndex,
-			"adopt_hash", fmt.Sprintf("%x", h.Hash[:8]),
-			"canonical_hash", fmt.Sprintf("%x", canonicalHash[:8]),
-		)
-		return nil
-	}
-
-	s.commitTransactionResultsLocked(stagedResults)
-	txResults := stagedResults.results
-	if adopted.IsValidated() {
-		s.enqueueValidatedHistoryPersist(adopted)
-	} else {
-		s.enqueueNodePersist(adopted)
-	}
-
-	// Rebuild openLedger only on forward adoption (backward-fills must not
-	// regress the open view).
-	if advanced {
-		openLedger, err := ledger.NewOpen(adopted, time.Now())
-		if err != nil {
-			return fmt.Errorf("failed to create open ledger: %w", err)
-		}
-		s.openLedger = openLedger
-		// Forward adopt = fresh start: rebuild via New on adopted.
-		if err := s.rebuildOpenLedgerViewLocked(); err != nil {
-			return err
-		}
-	}
-
-	// Forward adoption publishes catch-up ledgers. Below-tip history backfills
-	// remain silent so streams never run backward.
-	if advanced {
-		ledgerInfo := &LedgerInfo{
-			Sequence:  h.LedgerIndex,
-			Hash:      h.Hash,
-			CloseTime: adopted.CloseTime(),
-			Validated: adopted.IsValidated(),
-		}
-		// The event sink fires on *validated*, not *closed*; peer-adopt advances
-		// closedLedger only. Stash by hash for the next SetValidatedLedger to
-		// drain.
-		event := &LedgerAcceptedEvent{
-			LedgerInfo:         ledgerInfo,
-			Ledger:             adopted,
-			TransactionResults: txResults,
-		}
-		s.retainValidationCandidateLocked(adopted)
-		if s.hasEventSink() {
-			s.stashPendingValidationLocked(h.Hash, event)
-		}
-	}
-
-	s.logger.Info("Adopted ledger with full state from peer",
-		"seq", h.LedgerIndex,
-		"hash", fmt.Sprintf("%x", h.Hash[:8]),
-		"account_hash", fmt.Sprintf("%x", h.AccountHash[:8]),
-	)
-
-	// Cascade any held adoption awaiting this ledger (out-of-order replay-delta
-	// completions otherwise stall); also evicts stale held entries.
-	s.cascadeHeldAdoptionsLocked(ctx, adopted, cascadeDepth)
-
 	return nil
 }
