+++ internal/node/cleaner_integration_test.go
@@ -59,10 +59,6 @@ func TestLedgerCleanerProductionBoundaryRepairsDurableTransactionRoot(t *testing
 	env.Close()
 	closed := env.LastClosedLedger()
 	require.NotNil(t, closed)
-	stateMap, err := closed.StateMapSnapshot()
-	require.NoError(t, err)
-	txMap, err := closed.TxMapSnapshot()
-	require.NoError(t, err)
 	hdr := closed.Header()
 	require.NotZero(t, hdr.AccountHash)
 	require.NotZero(t, hdr.TxHash)
@@ -73,7 +69,7 @@ func TestLedgerCleanerProductionBoundaryRepairsDurableTransactionRoot(t *testing
 	}))
 	require.NotZero(t, transactionHash)
 
-	require.NoError(t, svc.AdoptLedgerWithState(ctx, &hdr, stateMap, txMap))
+	require.NoError(t, svc.SwitchToPreferredLedger(closed))
 	svc.SetValidatedLedger(hdr.LedgerIndex, hdr.Hash)
 	require.Equal(t, hdr.Hash, svc.GetValidatedLedger().Hash())
 	svc.FlushPersists()
